Moves CohortLookup source and header files to /src and /include
